Pixel

+++ assignment4/compute_temperatures_doubleglobal.cl
@@ -1,5 +1,5 @@
 __kernel void temperature_diffusion(
-    __global long double * told,
+    __global double * told,
     int l,
     int w,
     float dt
@@ -7,11 +7,11 @@ __kernel void temperature_diffusion(
 {
     int i = get_global_id(0);
     int j = get_global_id(1);
-    long double newtemperature;
+    double newtemperature;
     newtemperature = told[(i+1) * w + (j+1)] + (dt * 
                                 ((told[(i+1) * w + j] + told[(i+1) * w + (j+2)] + told[i * w + (j+1)] + told[(i+2) * w + (j+1)])/4 
                                 - told[(i+1) * w + (j+1)]));
-    
+
     barrier(CLK_GLOBAL_MEM_FENCE);
     told[(i+1) * w + (j+1)] = newtemperature;
 }
